normal_light

+++ build/build.sh
@@ -12,7 +12,9 @@ if [ ! -d buildcache ]; then
     mkdir buildcache
 fi
 
-rm -rf ${cache_path}
+if [ "$1" = "clean" ]; then
+    rm -rf ${cache_path}
+fi
 rm -rf ${output_path}
 rm -rf ${release_path}
